requirements.txt
endee
sentence-transformers
ollama
python-dotenv
scikit-learn
joblib
